+++ actix-web-lab/src/respond.rs
@@ -1,3 +1,9 @@
 //! Expiremental responders and response helpers.
 
-pub use crate::{csv::Csv, display_stream::DisplayStream, html::Html, ndjson::NdJson};
+pub use crate::{
+    csv::Csv,
+    display_stream::DisplayStream,
+    html::Html,
+    msgpack::{MessagePack, MessagePackNamed},
+    ndjson::NdJson,
+};